+++ src/main/kotlin/fr/enedis/grafana/dsl/dashboard/DashboardBuilder.kt
@@ -1,3 +1,10 @@
+/*
+ * SPDX-FileCopyrightText: 2021-2022 NBCO YooMoney LLC
+ *
+ * SPDX-License-Identifier: MIT
+ *
+ */
+
 package fr.enedis.grafana.dsl.dashboard
 
 import fr.enedis.grafana.dsl.DashboardElement