+++ src/main/kotlin/fr/enedis/grafana/dsl/dashboard/GraphTooltip.kt
@@ -1,3 +1,10 @@
+/*
+ * SPDX-FileCopyrightText: 2023-2025 Enedis
+ *
+ * SPDX-License-Identifier: MIT
+ *
+ */
+
 package fr.enedis.grafana.dsl.dashboard
 
 enum class GraphTooltip(val value: Int) {