+++ src/main/kotlin/fr/enedis/grafana/dsl/datasource/Elastic.kt
@@ -1,3 +1,10 @@
+/*
+ * SPDX-FileCopyrightText: 2023-2025 Enedis
+ *
+ * SPDX-License-Identifier: MIT
+ *
+ */
+
 import fr.enedis.grafana.dsl.datasource.Datasource
 
 /**