+++ src/main/kotlin/fr/enedis/grafana/dsl/metrics/ElasticMetric.kt
@@ -1,3 +1,10 @@
+/*
+ * SPDX-FileCopyrightText: 2023-2025 Enedis
+ *
+ * SPDX-License-Identifier: MIT
+ *
+ */
+
 package fr.enedis.grafana.dsl.metrics
 
 import org.json.JSONObject
@@ -356,4 +363,3 @@ class ElasticQueryMetricPipelineVariable constructor(
         }
     }
 }
-