+++ src/main/kotlin/fr/enedis/grafana/dsl/panels/barChart/StackingMode.kt
@@ -1,3 +1,10 @@
+/*
+ * SPDX-FileCopyrightText: 2023-2025 Enedis
+ *
+ * SPDX-License-Identifier: MIT
+ *
+ */
+
 package fr.enedis.grafana.dsl.panels.barChart
 
 enum class StackingMode(val value: String) {