+++ src/main/kotlin/fr/enedis/grafana/dsl/panels/barGauge/BarGaugePanel.kt
@@ -1,3 +1,10 @@
+/*
+ * SPDX-FileCopyrightText: 2023-2025 Enedis
+ *
+ * SPDX-License-Identifier: MIT
+ *
+ */
+
 package fr.enedis.grafana.dsl.panels.barGauge
 
 import fr.enedis.grafana.dsl.json.jsonObject
@@ -23,4 +30,3 @@ class BarGaugePanel(
     }
 
 }
-