+++ src/main/kotlin/fr/enedis/grafana/dsl/panels/table/Mapping.kt
@@ -1,3 +1,10 @@
+/*
+ * SPDX-FileCopyrightText: 2023-2025 Enedis
+ *
+ * SPDX-License-Identifier: MIT
+ *
+ */
+
 package fr.enedis.grafana.dsl.panels.table
 
 import org.json.JSONObject