+++ src/test/kotlin/fr/enedis/grafana/dsl/panels/alerting/AlertPanelBuilderTest.kt
@@ -1,3 +1,10 @@
+/*
+ * SPDX-FileCopyrightText: 2023-2025 Enedis
+ *
+ * SPDX-License-Identifier: MIT
+ *
+ */
+
 package fr.enedis.grafana.dsl.panels.alerting
 
 import fr.enedis.grafana.dsl.jsonFile