+++ src/main/kotlin/fr/enedis/grafana/dsl/panels/table/TablePanelFieldConfig.kt
@@ -7,6 +7,7 @@
 
 package fr.enedis.grafana.dsl.panels.table
 
+import fr.enedis.grafana.dsl.dashboard.link.DashboardLinks
 import fr.enedis.grafana.dsl.json.Json
 import fr.enedis.grafana.dsl.json.JsonArray
 import fr.enedis.grafana.dsl.json.jsonObject
@@ -16,8 +17,7 @@ import fr.enedis.grafana.dsl.panels.Thresholds
 import fr.enedis.grafana.dsl.panels.ThresholdsBuilder
 import fr.enedis.grafana.dsl.panels.fields.OverrideFieldConfig
 import fr.enedis.grafana.dsl.panels.fields.OverridesFieldConfigBuilder
-import fr.enedis.grafana.dsl.panels.table.Mapping
-import fr.enedis.grafana.dsl.panels.table.MappingsBuilder
+import fr.enedis.grafana.dsl.panels.stateTimeline.ColorScheme
 import org.json.JSONObject
 
 class TablePanelFieldConfig(
@@ -25,7 +25,10 @@ class TablePanelFieldConfig(
     private val thresholds: Thresholds = Thresholds(),
     private val mappings: List<Mapping> = emptyList(),
     private val overrides: List<OverrideFieldConfig> = emptyList(),
-    private val custom: TableCustomFieldConfig = TableCustomFieldConfig()
+    private val custom: TableCustomFieldConfig = TableCustomFieldConfig(),
+    private val links: DashboardLinks,
+    private val colorScheme: ColorScheme? = null,
+    private val noValue: String? = null,
 ) : Json<JSONObject> {
     override fun toJson(): JSONObject = jsonObject {
         "unit" to unit.unit
@@ -33,6 +36,9 @@ class TablePanelFieldConfig(
             "thresholds" to thresholds
             "mappings" to JsonArray(mappings)
             "custom" to custom
+            "links" to links
+            "color" to colorScheme
+            "noValue" to noValue
         }
         "overrides" to JsonArray(overrides)
     }
@@ -44,12 +50,18 @@ class TablePanelFieldConfigBuilder() {
     var mappings: List<Mapping> = emptyList()
     var overrides: List<OverrideFieldConfig> = emptyList()
     var custom: TableCustomFieldConfig = TableCustomFieldConfig()
+    var links: DashboardLinks = DashboardLinks(emptyList())
+    var colorScheme: ColorScheme? = null
+    var noValue: String? = null
     internal fun createTablePanelFieldConfig(): TablePanelFieldConfig = TablePanelFieldConfig(
-        unit,
-        thresholds,
-        mappings,
-        overrides,
-        custom
+        unit = unit,
+        thresholds = thresholds,
+        mappings = mappings,
+        overrides = overrides,
+        custom = custom,
+        links = links,
+        colorScheme = colorScheme,
+        noValue = noValue
     )
 
     fun thresholds(mode: ThresholdMode = ThresholdMode.ABSOLUTE, build: ThresholdsBuilder.() -> Unit) {