+++ changelog.md
@@ -1,3 +1,10 @@
+<!--
+  ~ SPDX-FileCopyrightText: 2023-2025 Enedis
+  ~
+  ~ SPDX-License-Identifier: MIT
+  ~
+  -->
+
 ## 1.3.7 (2024-12-16)
 - [[BUGFIX] Fix consistency with default values](dev/nexus/ccma/tools/grafana-dsl@378794e6683286071028699f4f0591a3d95beefb)
 - [[BUGFIX] Fix typo in json filename](dev/nexus/ccma/tools/grafana-dsl@e90bcc6356c8ae025d3da07dbf69d465262405a9)